services/orchestrator/requirements.txt
# cAIru Orchestrator Service Dependencies

# Redis client
redis>=5.0.0

# Database
aiosqlite>=0.19.0

# YAML for rules configuration
pyyaml>=6.0.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Scheduling for proactive rules
apscheduler>=3.10.0

# Observability
prometheus-client>=0.19.0
structlog>=23.2.0

